Roles Naprawa i front

+++ src/main/java/com/example/ztattom/entities/Role.java
@@ -1,5 +1,6 @@
 package com.example.ztattom.entities;
 
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -13,10 +14,10 @@ public class Role {
     @Id
     @Column(name = "role_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     public Role(){}
     public Role(String name) {this.name = name;}
-    public Role(Integer id){this.id = id;}
+    public Role(Long id){this.id = id;}
 
 }
